Revert class hierarchy changes
